+++ src/Pages/TaskPage/style.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+const StyledTaskPage = styled.div`
+    display: grid;
+    grid-template-areas: '. .';
+    grid-gap: 20px;
+`
+
+const StyledTaskPageList = styled.div`
+    display: grid;
+    grid-gap: 20px;
+`
+
+export { StyledTaskPage, StyledTaskPageList }
